fix(test): do not add a second `tests.conftest` to the pytest run
The socket-path fixture landed in partcad-service-json-rpc/tests/conftest.py and
broke every Pytest job, ubuntu included:

    ValueError: Plugin already registered under a different name:
      partcad-service-json-rpc/tests/conftest.py
      = <module 'tests.conftest' from 'partcad/tests/conftest.py'>

Both directories are packages named `tests`, so both conftest modules claim the
name `tests.conftest`, and pytest refuses the second one -- collection fails
before a single test runs. It only appears when one run collects both, which is
exactly what this branch introduced by adding partcad-service-json-rpc to the
matrix job; validating with `pytest partcad-service-json-rpc partcad-cli` could
not see it.

Define the fixture in the two modules that use it instead. A shared helper
module would not have been safer: with two `tests` packages on the path, any
name they share is ambiguous the same way.

Verified against CI's actual invocation this time -- `pytest partcad
partcad-cli partcad-service-json-rpc --collect-only` now collects 675 tests
where it previously raised.
